python/requirements.mcp.txt
# MCP Service Dependencies - Minimal
mcp==1.12.2
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
supabase==2.15.1
logfire>=0.30.0
fastapi>=0.104.0  # Required for mcp tools